refactor: deploy tool folder
