+++ src/tools/deploy/types.ts
@@ -1,0 +1,36 @@
+// Copyright Mia srl
+// SPDX-License-Identifier: Apache-2.0
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+export interface TriggerDeployResponse {
+  id: number
+  url: string
+}
+
+export interface PipelineStatus {
+  id: number
+  status: string
+}
+
+export interface Manifest {
+  content: string
+  name: string
+  resourceName: string
+  type: string
+}
+
+export interface CompareForDeployResponse {
+  lastDeployedManifests: Manifest[]
+  revisionManifests: Manifest[]
+}
